Move related notifier error to notifier-seeder package
